sikolbia-ml/requirements.txt
# Machine Learning Dependencies
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Data Processing
scipy==1.11.1
joblib==1.3.2

# Database Connection
mysql-connector-python==8.1.0
pymysql==1.1.0
SQLAlchemy==2.0.19

# API Integration
requests==2.31.0
flask==2.3.2
flask-cors==4.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Jupyter for development
jupyter==1.0.0
ipython==8.14.0

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
